created a get request to fetch leader board details

+++ controllers/habits-controller.js
@@ -101,9 +101,10 @@ const postHabit = async (req, res) => {
 
 const deleteHabit = async (req, res) => {
   const { id } = req.params;
+  const userId = req.user.id;
 
   try {
-    const rowsDeleted = await knex("habits").where({ id }).delete();
+    const rowsDeleted = await knex("habits").where({ id, user_id: userId }).delete();
 
     if (rowsDeleted === 0) {
       return res.status(404).json({ message: `Habit not found` });
